hegfysdsijfuj32432

+++ libft/ft_pow.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-long			ft_pow(double n, int pow)
+double			ft_pow(double n, double pow)
 {
 	return (pow ? n * ft_pow(n, pow - 1) : 1);
 }
